Simplified API calls🎉
The code behind API calls has been reformatted for easier understanding and use. All status codes are in a dictionary, to make it easier to know what they mean. Error messages do not get returned anymore. Since the status allows the frontend to know the error, it makes sense for localization purposes to have the frontend come up with a message.