Backend Image Serving🎉
The Gallery page has been started. This page can grab image data stored in the database, and request each image file. The backend loads the images and sends them in base64. Once the admin pages are complete, an admin will be able to control the images that show up in the gallery without modifying the website.