SKU Grouping
The shopping page now groups SKUs that have the same plant info. When a card is expanded, the user selects which size SKU they want before adding it to the cart. In the future, this grouping will be done on the backend, so less display images have to be passed