Image Fetching Update
Previously, there were many different functions for requesting images from the backend. Both by thumbnail and by full size. Now, images are requested by their id and a size label.

When an image is uploaded, its full size is stored. When there's a request for a smaller size that hasn't been made yet, the image is resized, stored as a new file, and returned