Auto password hashing
Password hashing and validation is now a child of the Objection Model class. This forces the password to be hashed right before a user is inserted into the database. It also allows an easy way to verify the password

Also, the route for setting a new pasword has been implemented