Magical query builder🧙🏻‍♂️
Created a query builder for dynamically fetching relationships from graphql info. For now, this only queries one level deep. Without this builder, the only way to fetch data for an object and its relationships was either to live with the N+1 problem, or manually convert a flat join table to a properly-formatted json object