CSS Sharing
Each file in ui used to be in its own folder. This was because they used to be accompanied by a styled.js. That was removed long ago, but the folder structure remained in case Material-UI's styling could have some use for it. After figuring out how to share styles with Material-UI - which is also in this commit - I decided to remove the extra folders.