Push notifications enabled manually
Moved push notification logic to sign up form and profile update page. When a user indicates that they'd like to receive notifications, the browser will automatically set it up. If they sign on with a new device and have notifications enabled, navigating to the profile page should (in theory) set it up automatically. Will need logic to do this on main page as well